+++ src/main/java/nextstep/app/SecurityConfig.java
@@ -5,10 +5,10 @@ import nextstep.app.domain.MemberRepository;
 import nextstep.oauth2.OAuth2AuthorizationRequestRedirectFilter;
 import nextstep.oauth2.OAuth2ClientProperties;
 import nextstep.oauth2.OAuth2LoginAuthenticationFilter;
-import nextstep.oauth2.client.ClientRegistration;
-import nextstep.oauth2.client.ClientRegistrationFactory;
-import nextstep.oauth2.client.ClientRegistrationRepository;
-import nextstep.oauth2.client.InMemoryClientRegistrationRepository;
+import nextstep.oauth2.registration.ClientRegistration;
+import nextstep.oauth2.registration.ClientRegistrationFactory;
+import nextstep.oauth2.registration.ClientRegistrationRepository;
+import nextstep.oauth2.registration.InMemoryClientRegistrationRepository;
 import nextstep.security.access.AnyRequestMatcher;
 import nextstep.security.access.MvcRequestMatcher;
 import nextstep.security.access.RequestMatcherEntry;
