+++ src/main/java/nextstep/oauth2/OAuth2AuthorizationRequestResolver.java
@@ -1,0 +1,16 @@
+package nextstep.oauth2;
+
+import jakarta.servlet.http.HttpServletRequest;
+import nextstep.oauth2.registration.ClientRegistrationRepository;
+
+public class OAuth2AuthorizationRequestResolver {
+    private final ClientRegistrationRepository clientRegistrationRepository;
+
+    public OAuth2AuthorizationRequestResolver(final ClientRegistrationRepository clientRegistrationRepository) {
+        this.clientRegistrationRepository = clientRegistrationRepository;
+    }
+
+    public OAuth2AuthorizationRequest resolve(HttpServletRequest request) {
+
+    }
+}
