+++ src/main/java/nextstep/oauth2/OAuth2LoginAuthenticationFilter.java
@@ -7,8 +7,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
 import nextstep.app.infrastructure.InmemoryMemberRepository;
-import nextstep.oauth2.client.ClientRegistration;
-import nextstep.oauth2.client.ClientRegistrationRepository;
+import nextstep.oauth2.registration.ClientRegistration;
+import nextstep.oauth2.registration.ClientRegistrationRepository;
 import nextstep.oauth2.exception.OAuth2RegistrationNotFoundException;
 import nextstep.oauth2.http.OAuth2ApiClient;
 import nextstep.oauth2.http.OAuth2User;
