+++ src/main/java/nextstep/oauth2/registration/ClientRegistration.java
@@ -1,4 +1,4 @@
-package nextstep.oauth2.client;
+package nextstep.oauth2.registration;
 
 import java.util.Collections;
 import java.util.Set;
