+++ src/main/java/nextstep/oauth2/registration/ClientRegistrationFactory.java
@@ -1,4 +1,4 @@
-package nextstep.oauth2.client;
+package nextstep.oauth2.registration;
 
 import nextstep.oauth2.OAuth2ClientProperties;
