+++ src/main/java/nextstep/oauth2/registration/ClientRegistrationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.oauth2.client;
+package nextstep.oauth2.registration;
 
 
 public interface ClientRegistrationRepository {
