+++ src/main/java/nextstep/oauth2/registration/InMemoryClientRegistrationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.oauth2.client;
+package nextstep.oauth2.registration;
 
 import java.util.Map;
